requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
werkzeug>=3.0.0
tenacity>=8.2.0
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
google-api-python-client>=2.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
